+++ core/src/main/java/com/moonwatch/core/model/ITokenValue.kt
@@ -1,9 +1,10 @@
 package com.moonwatch.core.model
 
+import java.math.BigDecimal
 import java.util.*
 
 interface ITokenValue {
   val address: String
-  val usd: Double
+  val usd: BigDecimal
   val updatedAt: Date
 }
