+++ core/src/main/java/com/moonwatch/core/repo/IAlertRepo.kt
@@ -1,9 +1,14 @@
 package com.moonwatch.core.repo
 
 import com.moonwatch.core.model.ITokenAlertWithValue
+import java.math.BigDecimal
 import kotlinx.coroutines.flow.Flow
 
 interface IAlertRepo {
   fun getTokenAlertsWithValue(): Flow<List<ITokenAlertWithValue>>
-  suspend fun addAlert(address: String, sellPriceTargetUsd: Double?, buyPriceTargetUsd: Double?)
+  suspend fun addAlert(
+      address: String,
+      sellPriceTargetUsd: BigDecimal?,
+      buyPriceTargetUsd: BigDecimal?
+  )
 }
