+++ core/src/main/java/com/moonwatch/core/usecase/AddAlert.kt
@@ -2,14 +2,15 @@ package com.moonwatch.core.usecase
 
 import com.moonwatch.core.repo.IAlertRepo
 import dagger.Reusable
+import java.math.BigDecimal
 import javax.inject.Inject
 
 @Reusable
 class AddAlert @Inject constructor(private val repo: IAlertRepo) {
   suspend operator fun invoke(
       address: String,
-      sellPriceTargetUsd: Double?,
-      buyPriceTargetUsd: Double?
+      sellPriceTargetUsd: BigDecimal?,
+      buyPriceTargetUsd: BigDecimal?
   ) {
     if (sellPriceTargetUsd == null && buyPriceTargetUsd == null) {
       throw IllegalArgumentException()
