+++ db/src/main/java/com/moonwatch/db/entity/TokenAlertEntity.kt
@@ -2,6 +2,7 @@ package com.moonwatch.db.entity
 
 import androidx.room.*
 import com.moonwatch.core.model.ITokenAlert
+import java.math.BigDecimal
 import java.util.*
 
 @Entity(
@@ -27,8 +28,8 @@ data class TokenAlertEntity(
     override val active: Boolean,
     @ColumnInfo(name = "created_at") override val createdAt: Date,
     @ColumnInfo(name = "last_fired_at") override val lastFiredAt: Date? = null,
-    @ColumnInfo(name = "sell_price_target_usd") override val sellPriceTargetUsd: Double? = null,
-    @ColumnInfo(name = "buy_price_target_usd") override val buyPriceTargetUsd: Double? = null,
+    @ColumnInfo(name = "sell_price_target_usd") override val sellPriceTargetUsd: BigDecimal? = null,
+    @ColumnInfo(name = "buy_price_target_usd") override val buyPriceTargetUsd: BigDecimal? = null,
 ) : ITokenAlert {
   @PrimaryKey(autoGenerate = true) override var id: Long = 0
 }
